+++ CHANGELOG.rst
@@ -16,7 +16,9 @@ CHANGELOG
 - Removed no_more_olga plugin. Now at https://github.com/shaunduncan/helga-no-more-olga (pypi helga-no-more-olga)
 - Removed oneliner plugin. Now at https://github.com/shaunduncan/helga-oneliner (pypi helga-onliner)
 - Removed poems plugin. Now at https://github.com/shaunduncan/helga-poems (pypi helga-poems)
+- Removed reminders plugin. Now at https://github.com/shaunduncan/helga-reminders (pypi helga-reminders)
 - Removed reviewboard plugin. Now at https://github.com/shaunduncan/helga-reviewboard (pypi helga-reviewboard)
+- Removed stfu plugin. Now at https://github.com/shaunduncan/helga-stfu (pypi helga-stfu)
 - Removed wiki_whois plugin. Now at https://github.com/shaunduncan/helga-wiki-whois (pypi helga-wiki-whois)